% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ius.R
\name{group_ius_according_to_mean_prevalence}
\alias{group_ius_according_to_mean_prevalence}
\title{Group Implementation Units (IUs) according to value of mean
prevalence estimated from data}
\usage{
group_ius_according_to_mean_prevalence(data, nsamples)
}
\arguments{
\item{data}{A dataframe containing a column named "Logit" and
another called "Sds" containing the statistical data. The
dataframe must have one row per IU}

\item{nsamples}{An integer describing the number of samples to use
when evaluating the mean prevalence from the logit and std
values.}
}
\value{
The same dataframe as \code{data} with an added column
containing the group number for each row (IU)
}
\description{
This functions reads in statistical prevalence data, in the form of
a logit and standard deviation value, for a range of IUs. For each
IU, it estimates the mean infection prevalence and determines in
which group the IU falls into.
}
